Using async await
