+++ Comments/ClientApp/package.json
@@ -29,6 +29,7 @@
     "babel-register": "^6.26.0",
     "chokidar-cli": "^1.2.0",
     "cross-env": "^5.1.3",
+    "es6-promise": "^4.2.4",
     "eslint": "^4.18.1",
     "eslint-config-react-app": "^2.1.0",
     "eslint-plugin-flowtype": "^2.45.0",
