+++ Comments/ClientApp/src/components/FetchData/FetchData.js
@@ -1,6 +1,6 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { fetchForecastData } from "./../actions/forecastActions";
+import { fetchForecastData } from "./forecastActions";
 
 // import fetch from "isomorphic-fetch";
 // todo: move to the forecast actions
@@ -19,17 +19,17 @@ class FetchData extends Component {
 					<tr>
 						<th>Date</th>
 						<th>Temp. (C)</th>
-						{/* <th>Temp. (F)</th>
-						<th>Summary</th> */}
+						<th>Temp. (F)</th>
+						<th>Summary</th>
 					</tr>
 				</thead>
 				<tbody>
 					{this.props.forecastData.map(forecast => (
-						<tr key={forecast.id}>
-							<td>{forecast.id}</td>
-							<td>{forecast.title}</td>
-							{/* <td>{forecast.body}</td>
-							<td>{forecast.summary}</td> */}
+						<tr key={forecast.dateFormatted}>
+							<td>{forecast.dateFormatted}</td>
+							<td>{forecast.temperatureC}</td>
+							<td>{forecast.temperatureF}</td>
+							<td>{forecast.summary}</td>
 						</tr>
 					))}
 				</tbody>
