+++ Comments/ClientApp/src/components/FetchData/forecastReducer.js
