--- Comments/ClientApp/src/components/get_forecasts.js
@@ -1,7 +1,0 @@
-const fetchData = async () => {
-	const response = await fetch("https://jsonplaceholder.typicode.com/posts/");
-	const data = await response.json();
-	return data;
-};
-
-export default fetchData;
